backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-tinymce==4.0.0
psycopg[binary]==3.2.1
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
